+++ microdata/tests.py
@@ -7,10 +7,19 @@ from lxml import html
 
 from os.path import dirname, join
 
+from pelican import Pelican
 from pelican.readers import Readers
 from pelican.settings import DEFAULT_CONFIG
 
 RESOURCES_PATH = join(dirname(__file__), 'test-resources')
+
+
+def get_settings(**kwargs):
+    settings = DEFAULT_CONFIG.copy()
+    for key, value in kwargs.items():
+        settings[key] = value
+    settings['PLUGINS'] = ['microdata']
+    return settings
 
 
 def normalize(text):
@@ -18,14 +27,9 @@ def normalize(text):
 
 
 class TestMicrodata(unittest.TestCase):
-    def setUp(self):
-        super(TestMicrodata, self).setUp()
-
-        import microdata
-        microdata.register()
-
-    def assert_rst_equal(self, rstfile, expected):
-        reader = Readers(settings=DEFAULT_CONFIG)
+    def assert_rst_equal(self, rstfile, expected, **kwargs):
+        pelican = Pelican(settings=get_settings(**kwargs))
+        reader = Readers(pelican.settings)
         content = reader.read_file(base_path=RESOURCES_PATH, path=rstfile).content
         self.assertEqual(normalize(content), normalize(expected))
 
@@ -43,7 +47,7 @@ class TestMicrodata(unittest.TestCase):
 
     def test_itemscope(self):
         expected = (
-            '<div itemscope itemtype="http://data-vocabulary.org/Person">'
+            '<div itemscope itemtype="http://schema.org/Person">'
             'My name is <span itemprop="name">John Doe</span>'
             '</div>'
         )
@@ -51,7 +55,7 @@ class TestMicrodata(unittest.TestCase):
 
     def test_itemscope_tag(self):
         expected = (
-            '<p itemscope itemtype="http://data-vocabulary.org/Person">'
+            '<p itemscope itemtype="http://schema.org/Person">'
             'My name is <span itemprop="name">John Doe</span>'
             '</p>'
         )
@@ -59,11 +63,11 @@ class TestMicrodata(unittest.TestCase):
 
     def test_nested_scope(self):
         expected = (
-            '<div itemscope itemtype="http://data-vocabulary.org/Person">'
+            '<div itemscope itemtype="http://schema.org/Person">'
             '<p>'
             'My name is <span itemprop="name">John Doe</span>'
             '</p>'
-            '<p itemprop="address" itemscope itemtype="http://data-vocabulary.org/Address">'
+            '<p itemprop="address" itemscope itemtype="http://schema.org/Address">'
             'My name is <span itemprop="name">John Doe</span>'
             '</p>'
             '</div>'
@@ -72,11 +76,20 @@ class TestMicrodata(unittest.TestCase):
 
     def test_nested_scope_compact(self):
         expected = (
-            '<p itemscope itemtype="http://data-vocabulary.org/Person">'
+            '<p itemscope itemtype="http://schema.org/Person">'
             'My name is <span itemprop="name">John Doe</span>'
-            '<span itemprop="address" itemscope itemtype="http://data-vocabulary.org/Address">'
+            '<span itemprop="address" itemscope itemtype="http://schema.org/Address">'
             'My name is <span itemprop="name">John Doe</span>'
             '</span>'
             '</p>'
         )
         self.assert_rst_equal('microdata_itemscope_nested_compact.rst', expected)
+
+    def test_custom_vocabulary(self):
+        expected = (
+            '<div itemscope itemtype="http://data-vocabulary.org/Person">'
+            'My name is <span itemprop="name">John Doe</span>'
+            '</div>'
+        )
+        self.assert_rst_equal('microdata_itemscope.rst', expected,
+                              MICRODATA_VOCABULARY='http://data-vocabulary.org')