+++ CHANGELOG.rst
@@ -6,6 +6,8 @@ Current
 
 - Support for Pelican 2.7 and Python 3.3+
 - Drop support for Python 3.2
+- Use schema.org as default vocabulary prefix
+- Make vocabulary prefix configurable with ``MICRODATA_VOCABULARY``
 
 
 0.2 (2013-06-21)